Add qgisserver upload project as zip feature

+++ qgisserver/models.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
@@ -10,8 +11,7 @@ from giscube.model_mixins import MetadataModelMixin, ResourceModelMixin
 from giscube.models import Category, Server
 from giscube.tilecache.models_mixins import TileCacheModelMixin
 from giscube.validators import validate_options_json_format
-from qgisserver.utils import (deactivate_services, patch_qgis_project, unique_service_directory,
-                              update_external_service, url_slash_join)
+from qgisserver.utils import deactivate_services, unique_service_directory, update_external_service, url_slash_join
 
 
 SERVICE_VISIBILITY_CHOICES = [
@@ -40,6 +40,11 @@ def validate_integer_pair_list(value):
         validate_integer_pair(line)
 
 
+def project_unique_service_directory(instance, filename):
+    filename = os.path.join('mapdata', filename)
+    return unique_service_directory(instance, filename)
+
+
 class Service(TileCacheModelMixin, models.Model):
     category = models.ForeignKey(
         Category, null=True, blank=True, on_delete=models.SET_NULL,
@@ -48,7 +53,8 @@ class Service(TileCacheModelMixin, models.Model):
     title = models.CharField(_('title'), max_length=100, null=True, blank=True)
     description = models.TextField(_('description'), null=True, blank=True)
     keywords = models.CharField(_('keywords'), max_length=200, null=True, blank=True)
-    project_file = models.FileField(_('project file'), upload_to=unique_service_directory)
+    project_file = models.FileField(
+        _('project file'), upload_to=project_unique_service_directory)
     service_path = models.CharField(_('service path'), max_length=255)
     active = models.BooleanField(_('active'), default=True, help_text='Enable/disable usage')
     visibility = models.CharField(_('visibility'), max_length=10, default='private',
@@ -97,17 +103,12 @@ class Service(TileCacheModelMixin, models.Model):
         url = "%s%s" % (server_url, mapfile)
         return url
 
-    def save(self, *args, **kwargs):
-        super(Service, self).save(*args, **kwargs)
-        patch_qgis_project(self)
-        update_external_service.delay(self.pk)
-
-    def __str__(self):
-        return str(self.title or self.name)
-
     @property
     def anonymous_view(self):
         return not (self.visibility == 'private')
+
+    def __str__(self):
+        return str(self.title or self.name)
 
     class Meta:
         verbose_name = _('Service')
@@ -115,13 +116,14 @@ class Service(TileCacheModelMixin, models.Model):
 
 
 @receiver(models.signals.post_delete, sender=Service)
-def auto_delete_file_on_delete(sender, instance, **kwargs):
+def auto_delete_service_path_on_delete(sender, instance, **kwargs):
     """
     Delete the project file if the service is deleted
     """
-    if instance.project_file:
-        if os.path.isfile(instance.project_file.path):
-            os.remove(instance.project_file.path)
+    if instance.service_path:
+        path = os.path.join(settings.MEDIA_ROOT, instance.service_path)
+        if os.path.isdir(path):
+            shutil.rmtree(path)
 
 
 @receiver(models.signals.pre_save, sender=Service)
@@ -131,7 +133,6 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     """
     if not instance.pk:
         return False
-
     try:
         old_file = Service.objects.get(pk=instance.pk).project_file
     except Service.DoesNotExist:
@@ -139,7 +140,7 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
 
     new_file = instance.project_file
     if not old_file == new_file:
-        if os.path.isfile(old_file.path):
+        if os.path.exists(old_file.path) and os.path.isfile(old_file.path):
             os.remove(old_file.path)
